requirements.txt
flask==3.0.3
flask-login==0.6.3
python-dotenv==1.0.1
supabase==2.5.0
werkzeug==3.0.3